requirements.txt
streamlit==1.30.0
opencv-python-headless==4.9.0.80
mediapipe==0.10.9
numpy==1.26.4
Pillow==10.2.0

